--- src/pages/SignUp/index.tsx
@@ -1,3 +1,0 @@
-import { SignUp } from './SignUp'
-
-export { SignUp }
